requirement.txt
pandas
seaborn
numpy
matplotlib
scikit-learn
catboost
xgboost
dill
pyyaml
-e .